+++ src/main/java/com/back/boundedContext/payout/domain/PayoutItem.java
@@ -21,7 +21,7 @@ public class PayoutItem extends BaseIdAndTime {
 
     String relTypeCode;
 
-    private int relId;
+    private long relId;
 
     private LocalDateTime paymentDate;
